+++ Editor/Listeners/VoidEventListenerEditor.cs
@@ -1,0 +1,7 @@
+using UnityEditor;
+
+namespace UnderLogic.Channels.Listeners.Editor
+{
+    [CustomEditor(typeof(VoidEventListener), true)]
+    public class VoidEventListenerEditor : EventListenerEditor<VoidEventChannel> { }
+}