+++ Runtime/FloatEventChannel.cs
@@ -3,5 +3,8 @@ using UnityEngine;
 namespace UnderLogic.Channels
 {
     [CreateAssetMenu(menuName = "Channels/Float Event Channel")]
-    public class FloatEventChannel : EventChannel<float> { }
+    public class FloatEventChannel : EventChannel<float>
+    {
+        public void RaiseEvent(int value) => RaiseEvent((float)value);
+    }
 }