+++ Runtime/Listeners/FloatEventListener.cs
@@ -1,4 +1,7 @@
 namespace UnderLogic.Channels.Listeners
 {
-    public class FloatEventListener : EventListener<float, FloatEventChannel> { }
+    public class FloatEventListener : EventListener<float, FloatEventChannel>
+    {
+        public void RaiseEvent(int value) => RaiseEvent((float)value);
+    }
 }